Update README.md
The purpose of this tool development is based on the fact that many people in their daily work are repeatedly writing scripts in order to study an App, and build related web frameworks to expose the interfaces they need. This tool perfectly covers this need and can make users more Focus on analytics apps.